remove OpTransactionRequest
