+++ crates/rpc-types/src/lib.rs
@@ -16,7 +16,7 @@ mod receipt;
 pub use receipt::{L1BlockInfo, OpTransactionReceipt, OpTransactionReceiptFields};
 
 mod transaction;
-pub use transaction::{OpTransactionFields, OpTransactionRequest, Transaction};
+pub use transaction::{OpTransactionFields, Transaction};
 
 pub mod error;
 pub use error::InvalidInboxEntry;
