+++ crates/rpc-types/src/transaction.rs
@@ -7,9 +7,6 @@ use alloy_serde::OtherFields;
 use op_alloy_consensus::OpTxEnvelope;
 use serde::{Deserialize, Serialize};
 
-mod request;
-pub use request::OpTransactionRequest;
-
 /// OP Transaction type
 #[derive(
     Clone, Debug, PartialEq, Eq, Serialize, Deserialize, derive_more::Deref, derive_more::DerefMut,
